Reorganize trap tests
