+++ LibreNMS/DB/Eloquent.php
@@ -125,7 +125,7 @@ class Eloquent
     public static function DB()
     {
         // check if Laravel is booted
-        if (defined('LARAVEL_START') && class_exists('DB')) {
+        if (class_exists('\DB') && \DB::connection()) {
             return \DB::connection();
         }
