+++ tests/Feature/SnmpTraps/TrapTestCase.php
@@ -1,0 +1,40 @@
+<?php
+/**
+ * TrapTestCase.php
+ *
+ * -Description-
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ * @package    LibreNMS
+ * @link       http://librenms.org
+ * @copyright  2019 Tony Murray
+ * @author     Tony Murray <murraytony@gmail.com>
+ */
+
+namespace LibreNMS\Tests\Feature\SnmpTraps;
+
+use App\Models\Eventlog;
+use Illuminate\Foundation\Testing\DatabaseTransactions;
+use LibreNMS\Tests\LaravelTestCase;
+
+abstract class TrapTestCase extends LaravelTestCase
+{
+    use DatabaseTransactions;
+
+    protected function lastEventlogMessage()
+    {
+        return Eventlog::orderBy('datetime', 'DESC')->first()->message;
+    }
+}
